backend/pwc_env/requirements.txt
dotenv==0.9.9
httpcore==0.10.2
httpx==0.14.3
paperswithcode-client
Flask
werkzeug==2.2.3
gunicorn==23.0.0